Logout Feature added plus code refactoring for css
I added a logout feature. This can be accessed when the admin hovers on
his/her name on the navbar. This feature is not available on mobile yet.
I also did a bit of code refactoring for the scss files. Other parts of
the codebase were modified to make for a better user expereince